Move settings components to their own folder
