+++ src/components/scenes/ChangeMiningFeeScene.tsx
@@ -11,11 +11,11 @@ import { FeeOption } from '../../types/types'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { showError } from '../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
+import { SettingsRadioRow } from '../settings/SettingsRadioRow'
 import { Alert } from '../themed/Alert'
 import { MainButton } from '../themed/MainButton'
 import { OutlinedTextInput } from '../themed/OutlinedTextInput'
 import { SceneHeader } from '../themed/SceneHeader'
-import { SettingsRadioRow } from '../themed/SettingsRadioRow'
 
 interface OwnProps extends EdgeSceneProps<'changeMiningFee'> {}
