+++ src/components/scenes/CurrencySettingsScene.tsx
@@ -8,9 +8,9 @@ import { useDispatch, useSelector } from '../../types/reactRedux'
 import { EdgeSceneProps } from '../../types/routerTypes'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
+import { SettingsHeaderRow } from '../settings/SettingsHeaderRow'
+import { SettingsRadioRow } from '../settings/SettingsRadioRow'
 import { MaybeBlockbookSetting, MaybeCustomServersSetting, MaybeElectrumSetting } from '../themed/MaybeCustomServersSetting'
-import { SettingsHeaderRow } from '../themed/SettingsHeaderRow'
-import { SettingsRadioRow } from '../themed/SettingsRadioRow'
 
 interface Props extends EdgeSceneProps<'currencySettings'> {}
